+++ src/main/java/nl/dtls/fairdatapoint/service/profile/ProfileService.java
@@ -22,6 +22,9 @@
  */
 package nl.dtls.fairdatapoint.service.profile;
 
+import nl.dtls.fairdatapoint.database.mongo.repository.ShapeRepository;
+import nl.dtls.fairdatapoint.entity.resource.ResourceDefinition;
+import nl.dtls.fairdatapoint.service.resource.ResourceDefinitionService;
 import nl.dtls.fairdatapoint.service.shape.ShapeService;
 import org.eclipse.rdf4j.model.IRI;
 import org.eclipse.rdf4j.model.Model;
@@ -51,32 +54,40 @@ public class ProfileService {
     private String persistentUrl;
 
     @Autowired
-    private ShapeService shapeService;
+    private ShapeRepository shapeRepository;
 
-    public Optional<Model> getProfileByUuid(String uuid, IRI uri) {
-        return shapeService.getShapeByUuid(uuid)
-                .map(shape -> {
-                    ModelBuilder modelBuilder = new ModelBuilder();
-                    Resource resource = bn();
-                    modelBuilder.subject(resource);
-                    modelBuilder.add(RDF.TYPE, i(format("%s#ResourceDescriptor", PROFILE_PREFIX)));
-                    modelBuilder.add(DCTERMS.FORMAT, i("https://w3id.org/mediatype/text/turtle"));
-                    modelBuilder.add(DCTERMS.CONFORMS_TO, i("https://www.w3.org/TR/shacl/"));
-                    modelBuilder.add(i(format("%s#hasRole", PROFILE_PREFIX)), i(format("%s/role#Validation",
-                            PROFILE_PREFIX)));
-                    modelBuilder.add(i(format("%s#hasArtifact", PROFILE_PREFIX)), i(format("%s/shapes/%s",
-                            persistentUrl, uuid)));
+    @Autowired
+    private ResourceDefinitionService resourceDefinitionService;
 
-                    Model profile = new LinkedHashModel();
-                    profile.add(uri, RDF.TYPE, i(format("%s#Profile", PROFILE_PREFIX)));
-                    profile.add(uri, RDFS.LABEL, l(format("Profile for %s", shape.getName())));
-                    profile.add(uri, i(format("%s#isProfileOf", PROFILE_PREFIX)), i(format("%s/profile/core",
-                            persistentUrl)));
-                    profile.add(uri, i(format("%s#hasResource", PROFILE_PREFIX)), resource);
-                    profile.addAll(new ArrayList<>(modelBuilder.build()));
-
-                    return profile;
-                });
+    private Model getProfileForResourceDefinition(ResourceDefinition rd, IRI uri) {
+        Model profile = new LinkedHashModel();
+        profile.add(uri, RDF.TYPE, i(format("%s#Profile", PROFILE_PREFIX)));
+        profile.add(uri, RDFS.LABEL, l(format("%s Profile", rd.getName())));
+        profile.add(uri, i(format("%s#isProfileOf", PROFILE_PREFIX)), i(format("%s/profile/core",
+                persistentUrl)));
+        rd.getShapeUuids().forEach(shapeUuid -> shapeRepository.findByUuid(shapeUuid).map(shape -> {
+            ModelBuilder modelBuilder = new ModelBuilder();
+            Resource resource = bn();
+            modelBuilder.subject(resource);
+            modelBuilder.add(RDF.TYPE, i(format("%s#ResourceDescriptor", PROFILE_PREFIX)));
+            modelBuilder.add(DCTERMS.FORMAT, i("https://w3id.org/mediatype/text/turtle"));
+            modelBuilder.add(DCTERMS.CONFORMS_TO, i("https://www.w3.org/TR/shacl/"));
+            modelBuilder.add(i(format("%s#hasRole", PROFILE_PREFIX)), i(format("%s/role#Validation",
+                    PROFILE_PREFIX)));
+            modelBuilder.add(i(format("%s#hasArtifact", PROFILE_PREFIX)), i(format("%s/shapes/%s",
+                    persistentUrl, shapeUuid)));
+            profile.add(uri, i(format("%s#hasResource", PROFILE_PREFIX)), resource);
+            profile.addAll(new ArrayList<>(modelBuilder.build()));
+            return null;
+        }));
+        return profile;
     }
 
+    public Optional<Model> getProfileByUuid(String uuid, IRI uri) {
+        return resourceDefinitionService.getByUuid(uuid).map(rd -> getProfileForResourceDefinition(rd, uri));
+    }
+
+    public IRI getProfileUri(ResourceDefinition rd) {
+        return i(format("%s/profile/%s", persistentUrl, rd.getUuid()));
+    }
 }